# CRM/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CombatDen API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/waivers/": {
      "get": {
        "tags": [
          "waivers"
        ],
        "summary": "List waivers for a gym",
        "description": "Lists non-deleted waivers for the gym, each with its current version number and the count of members who signed the current version.",
        "operationId": "list_waivers_api_v1_waivers__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Waivers listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverResponse"
                  },
                  "title": "Response List Waivers Api V1 Waivers  Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "waivers"
        ],
        "summary": "Create a waiver",
        "description": "Creates a waiver and publishes its first version.",
        "operationId": "create_waiver_api_v1_waivers__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaiverCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Waiver created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaiverResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "waivers"
        ],
        "summary": "Update a waiver",
        "description": "Renames a waiver and/or publishes a new version of its text. Editing the body publishes a new immutable version; an unchanged body is a no-op. Members must re-sign a newly published version.",
        "operationId": "update_waiver_api_v1_waivers__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaiverUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Waiver updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaiverResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Waiver not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "waivers"
        ],
        "summary": "Archive a waiver",
        "description": "Archives a waiver (soft-delete). Its versions and signatures are retained for the legal record.",
        "operationId": "delete_waiver_api_v1_waivers__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waiver_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waiver Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Waiver archived successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Waiver not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waivers/signatures/by-member/{member_id}": {
      "get": {
        "tags": [
          "waivers"
        ],
        "summary": "List a member's waiver status",
        "description": "Lists every non-deleted gym waiver with this member's latest sign status for each (for the member-detail Waivers section).",
        "operationId": "list_member_waiver_status_api_v1_waivers_signatures_by_member__member_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member waiver status listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberWaiverStatusRow"
                  },
                  "title": "Response List Member Waiver Status Api V1 Waivers Signatures By Member  Member Id  Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waivers/{waiver_id}/versions": {
      "get": {
        "tags": [
          "waivers"
        ],
        "summary": "List a waiver's versions",
        "description": "Lists a waiver's immutable version history, newest first, each with the count of members who signed that exact version.",
        "operationId": "list_waiver_versions_api_v1_waivers__waiver_id__versions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waiver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waiver Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Versions listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverVersionResponse"
                  },
                  "title": "Response List Waiver Versions Api V1 Waivers  Waiver Id  Versions Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waivers/{waiver_id}/signatures": {
      "get": {
        "tags": [
          "waivers"
        ],
        "summary": "List a waiver's signature roster",
        "description": "Lists every gym member and their latest sign status for this waiver \u2014 signed?, which version, when, and whether it is the current version.",
        "operationId": "list_waiver_signatories_api_v1_waivers__waiver_id__signatures_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waiver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waiver Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Roster listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverSignatoryRow"
                  },
                  "title": "Response List Waiver Signatories Api V1 Waivers  Waiver Id  Signatures Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waivers/{waiver_id}": {
      "get": {
        "tags": [
          "waivers"
        ],
        "summary": "Get a waiver",
        "description": "Fetches a single waiver with its current version body.",
        "operationId": "get_waiver_api_v1_waivers__waiver_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waiver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waiver Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Waiver fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaiverResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Waiver not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberWaiverStatusRow": {
        "properties": {
          "waiver_id": {
            "type": "string",
            "format": "uuid",
            "title": "Waiver Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "current_version_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Number"
          },
          "signed": {
            "type": "boolean",
            "title": "Signed"
          },
          "signed_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed Version Id"
          },
          "signed_version_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed Version Number"
          },
          "signed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed At"
          },
          "signed_current_version": {
            "type": "boolean",
            "title": "Signed Current Version",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "waiver_id",
          "name",
          "signed"
        ],
        "title": "MemberWaiverStatusRow",
        "description": "One row of a member's waiver status: a gym waiver + this member's latest\nsign status for it (for the member-detail Waivers section)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WaiverCreateRequest": {
        "properties": {
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "gym_id",
          "name",
          "body"
        ],
        "title": "WaiverCreateRequest",
        "description": "Create a waiver. Creating it also publishes version 1 from `body`."
      },
      "WaiverResponse": {
        "properties": {
          "waiver_id": {
            "type": "string",
            "format": "uuid",
            "title": "Waiver Id"
          },
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "current_version_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Number"
          },
          "current_version_signed_count": {
            "type": "integer",
            "title": "Current Version Signed Count",
            "default": 0
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "current_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WaiverVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "waiver_id",
          "gym_id",
          "name",
          "is_deleted",
          "created_at",
          "updated_at"
        ],
        "title": "WaiverResponse",
        "description": "A waiver catalog entry. `current_version` is populated by the single-get;\nthe list endpoint leaves it None (no body) and fills the summary counts."
      },
...
      "WaiverUpdateRequest": {
        "properties": {
          "waiver_id": {
            "type": "string",
            "format": "uuid",
            "title": "Waiver Id"
          },
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "data": {
            "$ref": "#/components/schemas/WaiverUpdateData"
          }
        },
        "type": "object",
        "required": [
          "waiver_id",
          "gym_id",
          "data"
        ],
        "title": "WaiverUpdateRequest",
        "description": "Rename a waiver and/or publish a new version of its text."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}